+++ src/hedera/hederaTestnetInfo.ts
@@ -2,12 +2,11 @@ import { EdgeCurrencyInfo } from 'edge-core-js/types'
 
 import { makeOuterPlugin } from '../common/innerPlugin'
 import type { HederaTools } from './HederaTools'
-import { HederaNetworkInfo, hederaOtherMethodNames } from './hederaTypes'
+import { HederaNetworkInfo } from './hederaTypes'
 
 const networkInfo: HederaNetworkInfo = {
-  creatorApiServers: ['https://creator.testnet.myhbarwallet.com'],
   mirrorNodes: ['https://testnet.mirrornode.hedera.com'],
-  client: 'TestNet',
+  client: 'testnet',
   checksumNetworkID: '1',
   maxFee: 900000
 }
@@ -40,7 +39,6 @@ const currencyInfo: EdgeCurrencyInfo = {
 export const hederatestnet = makeOuterPlugin<HederaNetworkInfo, HederaTools>({
   currencyInfo,
   networkInfo,
-  otherMethodNames: hederaOtherMethodNames,
 
   async getInnerPlugin() {
     return await import('./HederaTools')